server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.12",
    "google>=3.0.0",
    "google-genai>=1.17.0",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
]
